refactor: extract CLI recipe dry-run helpers
